fix to reflect multiple ids per sequence
When mapping sequences to nodes, the code seemed to assume that each
sequence was given a single id (from the header). It's possible this was
the case, but I think it was changed a while ago to reflect the fact
that each sequence can have more than one id (from different databases).

NOTE: this also highlighted an upstream bug which should be fixed by
biojs-io-fasta#3